Add working JFF FA loading :)
